Refactor parser structure
Changes made:
- Split parser file
- Add parser_expression file
- Add parser_statement file
- Add description to parser functions
